+++ opt_einsum_fx/__init__.py
@@ -1,7 +1,10 @@
+__version__ = "0.1.4"
+
 from ._script import jitable
 from ._opt_ein import optimize_einsums, optimize_einsums_full
 from ._fuse import fuse_einsums, fuse_scalars
 from ._cutensor import is_cuTENSOR_available
+from ._efficient_shape_prop import EfficientShapeProp
 
 __all__ = [
     "jitable",
@@ -10,4 +13,5 @@ __all__ = [
     "fuse_einsums",
     "fuse_scalars",
     "is_cuTENSOR_available",
+    "EfficientShapeProp",
 ]