+++ opt_einsum_fx/_opt_ein.py
@@ -3,7 +3,7 @@ import warnings
 
 import torch
 from torch import fx
-from torch.fx.passes.shape_prop import ShapeProp
+from ._efficient_shape_prop import EfficientShapeProp as ShapeProp
 
 import opt_einsum
 from opt_einsum.contract import _core_contract
@@ -117,6 +117,7 @@ def optimize_einsums(
         )
 
     new_graph = fx.Graph()
+    tracer = fx.proxy.GraphAppendingTracer(new_graph)
     # env keeps track of new injected nodes in addition to existing ones,
     # making sure they get into new_graph
     env = {}
@@ -149,7 +150,7 @@ def optimize_einsums(
                 # we can dispatch to opt_einsum and implicitly
                 # add it to the Graph by symbolically tracing it.
                 proxy_args = [
-                    fx.Proxy(env[x.name]) if isinstance(x, fx.Node) else x
+                    fx.Proxy(env[x.name], tracer=tracer) if isinstance(x, fx.Node) else x
                     for x in node.args
                 ]
                 # Use _core_contract to avoid `len()` calls that